Update monte_carlo.py
Created a loop to instantiate multiple instances from a string (here manual input, but will be provided by the fasta reading function later)

The Random method works (nearly) well. Only problem is if all random choices are taken, then an infinite loop is created. Should just add an iteration counter and stop if the loop do more than x time. Not perfect fix, but should prevent infinite loop lock.

The class was improved, with extended use of the metaclass put in place in the last commit. The code is now much more readable.